Divide Sidebar.tsx into NodeDisplay and EdgeDisplay files
